修复了admin端查不到岗位列表的bug

+++ src/main/java/com/dhu/Pojo/StudentQueryParam.java
@@ -11,7 +11,7 @@ public class StudentQueryParam {
     private Integer page = 1;
     private Integer pageSize = 10;
 
-    private Long studentId;
+    private Integer studentId;
     private String name;
     private String gender;
     private Long majorId;
